+++ modules/client/test/mocks/ethers-providers.ts
@@ -1,8 +1,0 @@
-const original = jest.requireActual("@ethersproject/providers");
-export const mockedEthersProviders = {
-  ...original,
-  getNetwork: jest.fn().mockReturnValue({ chainId: 31337, name: "goerli" }),
-};
-jest.mock("@ethersproject/providers", () => {
-  return mockedEthersProviders;
-});