+++ modules/common/src/errors.ts
@@ -156,7 +156,7 @@ export class DataDecodingError extends Error {
 
 export class EnsUnsuportedByNetworkError extends Error {
   constructor(network: string) {
-    super("Network " + network + " dows not support ens");
+    super("Network " + network + " does not support ens");
   }
 }
 export class InvalidContractAddressError extends Error {